refactor: products-detail-page/components에서 sections 디렉토리 분리
