+++ src/pages/products-detail-page/sections/ProductDetailSection.jsx
